style(test): preserve gui contract evidence anchor
